Add default-on detection and skipping of probable SSH sessions
This is not a security, but a convenience feature, and thus the
detection is based entirely on the presence of environmental variables
to indicate a *possible*/*probable* SSH session.

This allows effective skipping of Touch ID when making use of SSH. This
behavior can be disabled by adding the `allow_ssh` option to the entry
in `/etc/pam.d/sudo`.