Update build instructions
This change updates the build instructions to use the more modern CMake
build commands. These commands do not depend on the build generator
used.

Furthermore, this change also adds instructions for when `libpam`
cannot be found by CMake automatically. In that case, the user may need
to specify the prefix path manually to point CMake in the right
direction.